changed admin forms

+++ api/commerce/qna/views.py
@@ -1,4 +1,4 @@
-from rest_framework.decorators import api_view, permission_classes
+from rest_framework.decorators import api_view
 from .serializers import QNAListSerializer, QNACreateSerializer, \
     QNADeleteSerializer, QNACountSerializer, QNARetrieveSerializer
 from rest_framework.permissions import AllowAny, IsAuthenticated
@@ -8,7 +8,6 @@ from rest_framework import status
 
 
 @api_view(["GET"])
-@permission_classes([AllowAny])
 def qna_count(request, *args, **kwargs):
     clayful_review_client = ClayfulQNAClient()
     response = clayful_review_client.qna_count(product=kwargs['product'])
@@ -16,7 +15,6 @@ def qna_count(request, *args, **kwargs):
 
 
 @api_view(["GET"])
-@permission_classes([AllowAny])
 def qna_list(request, *args, **kwargs):
     clayful_review_client = ClayfulQNAClient()
     response = clayful_review_client.qna_list(product=kwargs['product'], page=kwargs['page'])
@@ -24,16 +22,18 @@ def qna_list(request, *args, **kwargs):
 
 
 @api_view(["GET"])
-@permission_classes([AllowAny])
 def get_qna(request, *args, **kwargs):
+    if not request.user.is_authenticated:
+        return Response({'error_msg': '로그인 후 이용해주세요,'}, status=status.HTTP_400_BAD_REQUEST)
     clayful_review_client = ClayfulQNAClient()
     response = clayful_review_client.get_qna(review_id=kwargs['review_id'])
     return Response(response.data, status=status.HTTP_200_OK)
 
 
 @api_view(["POST"])
-@permission_classes([IsAuthenticated])
 def create_qna(request, *args, **kwargs):
+    if not request.user.is_authenticated:
+        return Response({'error_msg': '로그인 후 이용해주세요,'}, status=status.HTTP_400_BAD_REQUEST)
     clayful_review_client = ClayfulQNAClient()
     response = clayful_review_client.create_qna(
         customer=request.header['clayful'],
@@ -48,8 +48,9 @@ def create_qna(request, *args, **kwargs):
 
 
 @api_view(["DELETE"])
-@permission_classes([IsAuthenticated])
 def delete_qna(request, *args, **kwargs):
+    if not request.user.is_authenticated:
+        return Response({'error_msg': '로그인 후 이용해주세요,'}, status=status.HTTP_400_BAD_REQUEST)
     clayful_review_client = ClayfulQNAClient()
     response = clayful_review_client.delete_qna(
         customer=request.header['clayful'],
